[upstream] Optimizations (erincatto/box2d#1050)
Improved performance, up to 20% gain in stacking scenarios.
Behavior change: contact solver bias stage no longer applies friction
(almost free perf gain)
Behavior change: b2World_Step still performs all operations, including
collision when timeStep == 0.
b2WorldDef now has b2Capacity that lets you pre-size some arrays to
reduce allocations.
Optimized hit events. These are basically free now.
Cleaned up and optimized solver scheduling.
Converted to new dynamic array type.
Added threading logic to prevent theoretical solver deadlock if the user
task system stalls worker 0.
Added Compounds benchmark.

+++ src/Box2D.NET/B2SolverStage.cs
@@ -14,7 +14,7 @@ namespace Box2D.NET
         public ArraySegment<B2SolverBlock> blocks;
         public int blockCount;
 
-        public int colorIndex;
+        public byte colorIndex;
 
         // todo consider false sharing of this atomic
         public B2AtomicInt completionCount;
